feat: Implement reply preview component
Introduces a `ReplyPreview` component to display information about the parent post in a reply thread. This enhances the context provided to users when viewing posts. The `PostBubble` component is updated to utilize this new component for displaying reply previews. The `PostContextIndicator` component is also refactored to better handle different context types.